Refactor common FS handlers in a separate header file.
This must be included by both fuse and nofuse source files.
